chore: [#183973645] consolidated web and api factories into shared, simple
