+++ api/src/api/licenseStatusRouter.test.ts
@@ -1,12 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
+import { generateLicenseStatusItem, generateNameAndAddress } from "@shared/test";
 import { Express } from "express";
 import request from "supertest";
-import {
-  generateLicenseData,
-  generateLicenseStatusItem,
-  generateNameAndAddress,
-  generateUserData,
-} from "../../test/factories";
+import { generateLicenseData, generateUserData } from "../../test/factories";
 import { setupExpress } from "../libs/express";
 import { licenseStatusRouterFactory } from "./licenseStatusRouter";
 import { getSignedInUserId } from "./userRouter";
