+++ api/src/api/taxFilingRouter.test.ts
@@ -1,7 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
+import { generateTaxIdAndBusinessName } from "@shared/test";
 import { Express } from "express";
 import request from "supertest";
-import { generateTaxFilingData, generateTaxIdAndBusinessName, generateUserData } from "../../test/factories";
+import { generateTaxFilingData, generateUserData } from "../../test/factories";
 import { EncryptionDecryptionClient, TaxFilingInterface, UserDataClient } from "../domain/types";
 import { setupExpress } from "../libs/express";
 import { taxFilingRouterFactory } from "./taxFilingRouter";
