+++ api/src/api/userRouter.test.ts
@@ -2,6 +2,9 @@ import { getCurrentDate, parseDate } from "@shared/dateHelpers";
 import { createEmptyFormationFormData } from "@shared/formationData";
 import {
   generateBusinessNameAvailability,
+  generateFormationData,
+  generateFormationSubmitResponse,
+  generateGetFilingResponse,
   getFirstAnnualFiling,
   getSecondAnnualFiling,
   getThirdAnnualFiling,
@@ -11,15 +14,13 @@ import { Express } from "express";
 import jwt from "jsonwebtoken";
 import request from "supertest";
 import {
-  generateFormationData,
-  generateFormationSubmitResponse,
-  generateGetFilingResponse,
   generateLicenseData,
   generateProfileData,
   generateTaxFilingData,
   generateUser,
   generateUserData,
 } from "../../test/factories";
+
 import { generateAnnualFilings, getLastCalledWith } from "../../test/helpers";
 import { EncryptionDecryptionClient, TimeStampBusinessSearch, UserDataClient } from "../domain/types";
 import { setupExpress } from "../libs/express";
