+++ api/src/domain/user/updateLicenseStatusFactory.test.ts
@@ -1,13 +1,7 @@
 import { getCurrentDate, parseDate } from "@shared/dateHelpers";
+import { generateLicenseStatusItem, generateLicenseStatusResult, generateNameAndAddress } from "@shared/test";
 import { UserData } from "@shared/userData";
-import {
-  generateLicenseData,
-  generateLicenseStatusItem,
-  generateLicenseStatusResult,
-  generateNameAndAddress,
-  generateProfileData,
-  generateUserData,
-} from "../../../test/factories";
+import { generateLicenseData, generateProfileData, generateUserData } from "../../../test/factories";
 import { UpdateLicenseStatus, UserDataClient } from "../types";
 import { updateLicenseStatusFactory } from "./updateLicenseStatusFactory";
