+++ shared/src/test/factories.ts
@@ -1,5 +1,18 @@
+import { getCurrentDateFormatted, getCurrentDateISOString } from "../dateHelpers";
+import { defaultDateFormat } from "../defaultConstants";
+import {
+  FormationData,
+  FormationLegalType,
+  FormationSubmitResponse,
+  GetFilingResponse,
+} from "../formationData";
 import { randomInt } from "../intHelpers";
+import { LicenseStatusItem, LicenseStatusResult, NameAndAddress } from "../license";
 import { Municipality } from "../municipality";
+import { arrayOfSectors, SectorType } from "../sector";
+import { TaxFiling, TaxFilingLookUpRequest } from "../taxFiling";
+import { Preferences } from "../userData";
+import { generateFormationFormData } from "./formationFactories";
 
 export const generateMunicipality = (overrides: Partial<Municipality>): Municipality => {
   return {
@@ -7,6 +20,113 @@ export const generateMunicipality = (overrides: Partial<Municipality>): Municipa
     name: `some-name-${randomInt()}`,
     county: `some-county-${randomInt()}`,
     id: `some-id-${randomInt()}`,
+    ...overrides,
+  };
+};
+
+export const generateFormationSubmitResponse = (
+  overrides: Partial<FormationSubmitResponse>
+): FormationSubmitResponse => {
+  return {
+    success: !!(randomInt() % 2),
+    token: `some-token-${randomInt()}`,
+    formationId: `some-id-${randomInt()}`,
+    redirect: `some-redirect-${randomInt()}`,
+    errors: [],
+    lastUpdatedISO: getCurrentDateISOString(),
+    ...overrides,
+  };
+};
+
+export const generateGetFilingResponse = (overrides: Partial<GetFilingResponse>): GetFilingResponse => {
+  return {
+    success: true,
+    entityId: `some-entity-${randomInt()}`,
+    transactionDate: getCurrentDateISOString(),
+    confirmationNumber: `some-confirmation-number-${randomInt()}`,
+    formationDoc: `some-formation-doc-${randomInt()}`,
+    standingDoc: `some-standing-doc-${randomInt()}`,
+    certifiedDoc: `some-certified-doc-${randomInt()}`,
+    ...overrides,
+  };
+};
+
+export const generateFormationData = (
+  overrides: Partial<FormationData>,
+  legalStructureId?: FormationLegalType
+): FormationData => {
+  return {
+    formationFormData: generateFormationFormData({}, { legalStructureId }),
+    businessNameAvailability: undefined,
+    formationResponse: undefined,
+    getFilingResponse: undefined,
+    completedFilingPayment: false,
+    lastVisitedPageIndex: 0,
+    ...overrides,
+  };
+};
+
+export const randomSector = (): SectorType => {
+  const randomIndex = Math.floor(Math.random() * arrayOfSectors.length);
+  return arrayOfSectors[randomIndex];
+};
+
+export const generatePreferences = (overrides: Partial<Preferences>): Preferences => {
+  return {
+    roadmapOpenSections: ["PLAN", "START"],
+    roadmapOpenSteps: [randomInt()],
+    hiddenFundingIds: [],
+    hiddenCertificationIds: [],
+    visibleSidebarCards: ["welcome"],
+    returnToLink: "",
+    isCalendarFullView: !(randomInt() % 2),
+    isHideableRoadmapOpen: !(randomInt() % 2),
+    phaseNewlyChanged: false,
+    ...overrides,
+  };
+};
+
+export const generateNameAndAddress = (overrides: Partial<NameAndAddress>): NameAndAddress => {
+  return {
+    name: `some-name-${randomInt()}`,
+    addressLine1: `some-address-1-${randomInt()}`,
+    addressLine2: `some-address-2-${randomInt()}`,
+    zipCode: `some-zipcode-${randomInt()}`,
+    ...overrides,
+  };
+};
+
+export const generateTaxIdAndBusinessName = (
+  overrides: Partial<TaxFilingLookUpRequest>
+): TaxFilingLookUpRequest => {
+  return {
+    businessName: `some-name-${randomInt()}`,
+    taxId: `${randomInt(12)}`,
+    encryptedTaxId: "some-encrypted-value",
+    ...overrides,
+  };
+};
+
+export const generateLicenseStatusResult = (overrides: Partial<LicenseStatusResult>): LicenseStatusResult => {
+  return {
+    status: "PENDING",
+    checklistItems: [generateLicenseStatusItem({})],
+    ...overrides,
+  };
+};
+
+export const generateLicenseStatusItem = (overrides: Partial<LicenseStatusItem>): LicenseStatusItem => {
+  return {
+    title: `some-title-${randomInt()}`,
+    status: "ACTIVE",
+    ...overrides,
+  };
+};
+
+export const generateTaxFiling = (overrides: Partial<TaxFiling>): TaxFiling => {
+  return {
+    identifier: `some-identifier-${randomInt()}`,
+    dueDate: getCurrentDateFormatted(defaultDateFormat),
     ...overrides,
   };
 };
