+++ web/src/components/tasks/business-formation/success/FormationSuccessPage.test.tsx
@@ -1,9 +1,10 @@
 import { FormationSuccessPage } from "@/components/tasks/business-formation/success/FormationSuccessPage";
 import { getMergedConfig } from "@/contexts/configContext";
-import { generateFormationData, generateGetFilingResponse, generateUserData } from "@/test/factories";
+import { generateUserData } from "@/test/factories";
 import { generateFormationProfileData } from "@/test/helpers/helpers-formation";
 import { setMockDocumentsResponse, useMockDocuments } from "@/test/mock/mockUseDocuments";
 import { FormationLegalType, GetFilingResponse, ProfileData } from "@businessnjgovnavigator/shared";
+import { generateFormationData, generateGetFilingResponse } from "@businessnjgovnavigator/shared/test";
 import { render, screen } from "@testing-library/react";
 
 jest.mock("@/lib/data-hooks/useDocuments");
