+++ web/src/lib/cms/previews/DashboardCalendarPreview.tsx
@@ -6,13 +6,12 @@ import { usePreviewConfig } from "@/lib/cms/helpers/usePreviewConfig";
 import { usePreviewRef } from "@/lib/cms/helpers/usePreviewRef";
 import {
   generateOperateReference,
-  generatePreferences,
   generateProfileData,
-  generateTaxFiling,
   generateTaxFilingData,
   generateUserData,
 } from "@/test/factories";
 import { defaultDateFormat } from "@businessnjgovnavigator/shared/index";
+import { generatePreferences, generateTaxFiling } from "@businessnjgovnavigator/shared/test";
 import { createTheme, ThemeProvider } from "@mui/material";
 import dayjs from "dayjs";
 import { ReactElement } from "react";
