+++ web/src/lib/data-hooks/useUpdateTaskProgress.test.tsx
@@ -3,19 +3,14 @@
 import { useUpdateTaskProgress } from "@/lib/data-hooks/useUpdateTaskProgress";
 import { useUserData } from "@/lib/data-hooks/useUserData";
 import { UpdateQueue } from "@/lib/types/types";
-import {
-  generatePreferences,
-  generateRoadmap,
-  generateStep,
-  generateTask,
-  generateUserData,
-} from "@/test/factories";
+import { generateRoadmap, generateStep, generateTask, generateUserData } from "@/test/factories";
 import { setMockRoadmapResponse, useMockRoadmap } from "@/test/mock/mockUseRoadmap";
 import {
   currentUserData,
   setupStatefulUserDataContext,
   WithStatefulUserData,
 } from "@/test/mock/withStatefulUserData";
+import { generatePreferences } from "@businessnjgovnavigator/shared/test";
 import { TaskProgress, UserData } from "@businessnjgovnavigator/shared/userData";
 import { act, render } from "@testing-library/react";
 import { ReactNode } from "react";
