+++ web/test/pages/onboarding/onboarding-owning.test.tsx
@@ -2,7 +2,6 @@ import { getMergedConfig } from "@/contexts/configContext";
 import * as api from "@/lib/api-client/apiClient";
 import { templateEval } from "@/lib/utils/helpers";
 import {
-  generatePreferences,
   generateProfileData,
   generateTaxFilingData,
   generateUser,
@@ -18,6 +17,7 @@ import {
   runSelfRegPageTests,
 } from "@/test/pages/onboarding/helpers-onboarding";
 import { createEmptyUserData, generateMunicipality, ProfileData } from "@businessnjgovnavigator/shared/";
+import { generatePreferences } from "@businessnjgovnavigator/shared/test";
 import { UserData } from "@businessnjgovnavigator/shared/userData";
 import { act, screen, waitFor, within } from "@testing-library/react";
